Added a mode slider + a few lines of code to see the difference map between the GT and the approximation.

+++ src/cpp/Shading.cpp
@@ -87,7 +87,7 @@ void Shading::configureProgram()
 
     glProgramUniform1i (Parameters::getInstance()->g_programs[PROGRAM_SHADING],
             Parameters::getInstance()->g_uniform_locations[LOCATION_SHADING_GROUNDTRUTH],
-            Parameters::getInstance()->g_ground_truth);
+            (int)Parameters::getInstance()->g_ground_truth);
 
     glProgramUniform1f(Parameters::getInstance()->g_programs[PROGRAM_SHADING],
                             Parameters::getInstance()->g_uniform_locations[LOCATION_SHADING_SIZETEX],
